refactor: extract insight cash flow chart
